requirements.txt
dlib
grpcio==1.18.0
grpcio-tools==1.18.0
h5py==2.8.0
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
numpy==1.16.0
opencv-python==3.4.3.18
pandas==0.23.4
progressbar==2.5
protobuf==3.6.1
scikit-learn==0.20.0
sklearn==0.0
tensorboard==1.10.0
tensorboardX==1.4
tensorflow==1.10.1

